refactor(library/ShimmedStabilityAIClient): aligns file structure with "SDXL" namespace
