+++ src/library/ShimmedStabilityAIClient/SDXL/DiffusionStepCount.ts
